+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 Changelog
 =========
-## Unreleased
+## 1.0.0-rc.11 - 2018-02-18
 ### Removed
 - RecordHelper::configure, RecordHelper::create and RecordHelper::populate as they're code smell
 - ElementHelper due to code smell